Rename baseload analysis class, mark methods private
