+++ app/services/baseload/baseload_meter_breakdown_service.rb
@@ -54,7 +54,7 @@ module Baseload
           meter.mpan_mprn,
           {
             kw: meter.amr_data.average_baseload_kw_date_range(sheffield_solar_pv: meter.sheffield_simulated_solar_pv_panels?),
-            £: ElectricityBaseloadAnalysis.new(meter).scaled_annual_baseload_cost_£(:£)
+            £: BaseloadCalculator.new(meter).scaled_annual_baseload_cost_£(:£)
           }
         ]
       end.to_h
