--- spec/lib/dashboard/modelling/electricity/electricity_baseload_analysis_spec.rb
@@ -1,30 +1,0 @@
-require 'spec_helper'
-
-describe ElectricityBaseloadAnalysis do
-
-  let(:day_count)    { 365 }
-  let(:amr_data)    { build(:amr_data, :with_days, day_count: day_count) }
-  let(:meter)       { build(:meter, amr_data: amr_data) }
-  let(:analyser)    { ElectricityBaseloadAnalysis.new(meter) }
-
-  context '#one_years_data?' do
-    it 'returns true with a years worth of data' do
-      expect(analyser.one_years_data?).to be true
-    end
-    context 'with limited data' do
-      let(:day_count)    { 30 }
-      it 'returns false' do
-        expect(analyser.one_years_data?).to be false
-      end
-    end
-    context 'with historical asof_date' do
-      let(:day_count)    { 365 * 2 }
-      it 'returns true if there is a years worth of data before then' do
-        expect(analyser.one_years_data?(Date.today - 365)).to be true
-      end
-      it 'returns false if not enough data' do
-        expect(analyser.one_years_data?(Date.today - 500)).to be false
-      end
-    end
-  end
-end
